Added the Random synchronizer, and fixed a variety of bugs detected during the debugging of RandomSync.
git-svn-id: https://svn.code.sf.net/p/tricklesync/code/trunk@64 dd4b9e15-e244-0410-b754-811dee3e6019

+++ SyncSimulationFramework/ProbabilityController.h
@@ -34,7 +34,7 @@
 @property(readonly) double arrivalRate;
 @property(readonly) unsigned int timeIntervalInSeconds;
 
-+ (double)bernoulliProbabilityPerSecondFromEventsPerHour:(int)ratePerHour;
++ (double)bernoulliProbabilityPerSecondFromEventsPerHour:(double)ratePerHour;
 + (double)factorial:(long)x;
 + (double)poissonProbabilityWithArrivalRate:(double)rate
                            withTimeInterval:(int)t
